New Branch for Controller Refactoring. Deets below
It makes a lot of sense to allow the base controller class to handle a
lot of the checking and initialization for us.  Based on that logic, it
makes sense to store pointers to what needs to be initialized in the
current class, and then call back to super for the leg work that then
needs to be done based on those.  Fairly simple setup using that model,
and should be rather well documented in code.  The solution for views
that pass in live data is also based on paramaters as well.

NOTE: Discussion with Dan and myself noted the use of a flush command
for in-page-actions vs global-actions (flushing the entire DOM E.g.
  project show or profile show).  There is still some debate on how to
  implementat that, so that is still ongoing.

This has been set up in a seperate branch to evaluate its usefulness,
and to discuss with team before merging into a sprint branch (if at all
needed).